+++ handlers/entity_handlers/__init__.py
@@ -1,0 +1,16 @@
+from .characters import CharactersConversationHandler, CharactersHandler
+from .comics import ComicsConversationHandler, ComicsHandler
+from .events import EventsConversationHandler, EventsHandler
+from .series import SeriesConversationHandler, SeriesHandler
+
+
+__all__ = [
+    "CharactersConversationHandler",
+    "CharactersHandler",
+    "ComicsConversationHandler",
+    "ComicsHandler",
+    "EventsConversationHandler",
+    "EventsHandler",
+    "SeriesConversationHandler",
+    "SeriesHandler",
+]
